refactor: move client state query cache ownership
